Added where query and a if __name__ for testing.
The get_all_from_where simplifies making a query on a table with where
clauses. where clauses are sent in on the form column=(cmp,var) where cmp
is the comparison to use (eg. =,<,>,!=,LIKE etc.), column is the database
column to be compared and var is the value it compares to.