Added tasks.py, threading module
Apparently had not commited this even though i've worked on it a bit.
Basically the module that will contain the functions to be threaded like
database calls, the gui widgets uses start_task to create a new thread
with a task. For this an add_func has to be supplied which is a method in
the widget class that handles how to show whatever is added on the queue.
This is accomplished by using the widgets on after method to ensure that
the add_func will be called by gui-thread and not the worker thread.